feature: picky / gallery 插件开发修复

+++ bot/src/extensions/utils.py
@@ -109,24 +109,25 @@ async def get_chat_image(url, file, path, img_type=ImageType.chat, timeout=0, re
 async def get_image(url, file, path, img_type, _proxies=None):
     img_sql, _ = Image.get_or_create(filename=file, type_id=img_type.value)
 
-    async with httpx.AsyncClient(
-            proxies=_proxies
-    ) as cli:
-        resp = await cli.get(url)
+    if not img_sql.file_existed:
+        async with httpx.AsyncClient(
+                proxies=_proxies
+        ) as cli:
+            resp = await cli.get(url)
 
-        if resp.status_code == 200:
-            img = resp.content
-            suffix = what(img)
-            if not os.path.exists(path):
-                print(f'创建路径 {path}')
-                os.makedirs(path)
-            with open(path / f"{file}.{suffix}", 'wb') as fi:
-                fi.write(img)
+            if resp.status_code == 200:
+                img = resp.content
+                suffix = what(img)
+                if not os.path.exists(path):
+                    print(f'创建路径 {path}')
+                    os.makedirs(path)
+                with open(path / f"{file}.{suffix}", 'wb') as fi:
+                    fi.write(img)
 
-            img_sql.suffix = suffix
-            img_sql.file_existed = True
-            img_sql.update_time = datetime.datetime.now()
-            img_sql.save()
+                img_sql.suffix = suffix
+                img_sql.file_existed = True
+                img_sql.update_time = datetime.datetime.now()
+                img_sql.save()
 
     return img_sql
